fix bug in dataset
