+++ .gitignore
@@ -3,3 +3,5 @@ DATASET_old/
 .idea/
 *.mp4
 *.pyc
+*.ipynb
+/temp
